% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_cdf.R
\name{gh.cdf}
\alias{gh.cdf}
\title{CDF of GH distribution
Eq. (3)}
\usage{
gh.cdf(
  x,
  mu = 0,
  alpha = 1,
  beta = 0,
  delta = 1,
  p = -0.5,
  param = c(mu, delta, alpha, beta, p),
  n.quad = 8,
  correct = T,
  lower.tail = T
)
}
\arguments{
\item{mu}{GH parameter in the paper}

\item{alpha}{GH parameter in the paper}

\item{beta}{GH parameter in the paper}

\item{delta}{GH parameter in the paper}

\item{p}{GH parameter in the paper}

\item{param}{(mu, delta, alpha, beta, p) instead of above}

\item{n.quad}{number of quadrature points}

\item{correct}{normalize weights if TRUE}
}
\description{
CDF of GH distribution
Eq. (3)
}
