% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_quad.R
\name{gh.quad}
\alias{gh.quad}
\title{Double quadrature of the GH distribution}
\usage{
gh.quad(
  x,
  mu = 0,
  alpha = 1,
  beta = 0,
  delta = 1,
  p = -0.5,
  param = c(mu, delta, alpha, beta, p),
  n.quad = c(8, 4),
  correct = T
)
}
\arguments{
\item{mu}{GH parameter in the paper}

\item{alpha}{GH parameter in the paper}

\item{beta}{GH parameter in the paper}

\item{delta}{GH parameter in the paper}

\item{p}{GH parameter in the paper}

\item{param}{(mu, delta, alpha, beta, p) instead of above}

\item{n.quad}{number of quadrature points}

\item{correct}{normalize weights if TRUE}
}
\description{
Double quadrature of the GH distribution
}
