% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_rand.R
\name{gh.rand}
\alias{gh.rand}
\title{GH random variate
Eq. (5)}
\usage{
gh.rand(
  n = 99,
  mu = 0,
  alpha = 1,
  beta = 0,
  delta = 1,
  p = -0.5,
  param = c(mu, delta, alpha, beta, p),
  n.quad = 50,
  correct = T,
  GIGrvg = F,
  antithetic = T
)
}
\arguments{
\item{mu}{GH parameter in the paper}

\item{alpha}{GH parameter in the paper}

\item{beta}{GH parameter in the paper}

\item{delta}{GH parameter in the paper}

\item{p}{GH parameter in the paper}

\item{param}{(mu, delta, alpha, beta, p) instead of above}

\item{n.quad}{number of quadrature points}

\item{correct}{normalize weights if TRUE}

\item{GIGrvg}{If TRUE, use GIGrvg::rgig instead}
}
\description{
GH random variate
Eq. (5)
}
