% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig_mgf_num.R
\name{gig.mgf.num}
\alias{gig.mgf.num}
\title{Moment generating function of GIG distribution
Numerically evaluated at t
Eq. (11)}
\usage{
gig.mgf.num(t, gamma = 1, delta = 1, p = -0.5, n.quad = 50, correct = F)
}
\arguments{
\item{t}{dummy variable value}

\item{gamma}{GIG parameter in the paper}

\item{delta}{GIG parameter in the paper}

\item{p}{GIG parameter in the paper}

\item{n.quad}{number of quadrature points}

\item{correct}{normalize weights if TRUE}
}
\description{
Moment generating function of GIG distribution
Numerically evaluated at t
Eq. (11)
}
