% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig_mgf.R
\name{gig.mgf}
\alias{gig.mgf}
\title{Moment generating function of GIG distribution
Eq. (10)}
\usage{
gig.mgf(t, gamma = 1, delta = 1, p = -0.5)
}
\arguments{
\item{t}{dummy variable value}

\item{gamma}{GIG parameter in the paper}

\item{delta}{GIG parameter in the paper}

\item{p}{GIG parameter in the paper}
}
\description{
Moment generating function of GIG distribution
Eq. (10)
}
