% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig_mom_num.R
\name{gig.mom.num}
\alias{gig.mom.num}
\title{Numerical moment of GIG distribution.}
\usage{
gig.mom.num(order = 2, gamma = 1, delta = 1, p = -0.5, n.quad = 8, correct = T)
}
\arguments{
\item{order}{moment order}

\item{gamma}{GIG parameter in the paper}

\item{delta}{GIG parameter in the paper}

\item{p}{GIG parameter in the paper}

\item{n.quad}{number of quadrature points}

\item{correct}{normalize weights if TRUE}
}
\description{
Numerical moment of GIG distribution.
}
