% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig_mom.R
\name{gig.mom}
\alias{gig.mom}
\title{Analytic moments of GIG distribution.
Eq. (9)}
\usage{
gig.mom(order = 2, gamma = 1, delta = 1, p = -0.5)
}
\arguments{
\item{order}{moment order}

\item{gamma}{GIG parameter in the paper}

\item{delta}{GIG parameter in the paper}

\item{p}{GIG parameter in the paper}
}
\description{
Analytic moments of GIG distribution.
Eq. (9)
}
