% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig_quad.R
\name{gig.quad}
\alias{gig.quad}
\title{GIG quadrature
Corollary 1 in the paper}
\usage{
gig.quad(gamma = 1, delta = 1, p = -0.5, n.quad = 8, correct = T)
}
\arguments{
\item{gamma}{GIG parameter in the paper}

\item{delta}{GIG parameter in the paper}

\item{p}{GIG parameter in the paper}

\item{n.quad}{number of quadrature points}

\item{correct}{normalize weights if TRUE}
}
\description{
GIG quadrature
Corollary 1 in the paper
}
