% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig_rand.R
\name{gig.rand}
\alias{gig.rand}
\title{GIG random variate}
\usage{
gig.rand(
  n = 99,
  gamma = 1,
  delta = 1,
  p = -0.5,
  n.quad = 50,
  correct = T,
  antithetic = F
)
}
\arguments{
\item{n}{number of RVs}

\item{gamma}{GIG parameter in the paper}

\item{delta}{GIG parameter in the paper}

\item{p}{GIG parameter in the paper}

\item{n.quad}{number of quadrature points}

\item{correct}{normalize weights if TRUE}
}
\description{
GIG random variate
}
