% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_mom_num.R
\name{ig.mom.num}
\alias{ig.mom.num}
\title{IG moment numerically evaluated with IG quadrature}
\usage{
ig.mom.num(order = 2, mu = 1, lambda = 1, n.quad = 8)
}
\arguments{
\item{order}{moment order}

\item{mu}{IG parameter}

\item{lambda}{IG parameter}

\item{n.quad}{number of quadrature points}
}
\description{
IG moment numerically evaluated with IG quadrature
}
