% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_mom.R
\name{ig.mom}
\alias{ig.mom}
\title{IG moment}
\usage{
ig.mom(order = 2, mu = 1, lambda = 1)
}
\arguments{
\item{order}{moment order}

\item{mu}{IG parameter}

\item{lambda}{IG parameter}
}
\description{
IG moment
}
