% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_quad.R
\name{ig.quad}
\alias{ig.quad}
\title{Inverse Gaussian Quadrature
Theorem 1 in the paper}
\usage{
ig.quad(mu = 1, lambda = 1, n.quad = 8)
}
\arguments{
\item{mu}{IG parameter}

\item{lambda}{IG parameter}

\item{n.quad}{number of quadrature points}
}
\description{
Inverse Gaussian Quadrature
Theorem 1 in the paper
}
